[5.7.5.5] shift out key from rfs inode to prompt inodes capacity of rfs name node by chaoyong zhou on Jul 8,2018

+++ bgn_ngx/bgn_src/crfsmc.c
@@ -193,7 +193,7 @@ CRFSNP_FNODE *crfsmc_reserve_np_no_lock(CRFSMC *crfsmc, const CSTRING *file_path
         return (NULL_PTR);
     }
 
-    CRFSNP_ITEM_CREATE_TIME(crfsnp_item) = 0/*task_brd_default_get_time()*/;
+    CRFSNP_ITEM_CREATE_TIME(crfsnp_item) = (uint32_t)0/*task_brd_default_get_time()*/;
 
     if(do_log(SEC_0140_CRFSMC, 9))
     {
